interfaceの廃止、全てtypeにしてフォルダ、ファイルを分け直し
