+++ src/interfaces/index.ts
@@ -21,15 +21,4 @@ export interface User {
   allowPasswordChange: boolean
 }
 
-export type ShipData = {
-  id: number
-  regiNumber: string
-};
 
-export type ScheduleData = {
-  id: number
-  from: string
-  to: string
-  depTime: string
-  arrTime: string
-}
