+++ src/types/schedule.ts
@@ -1,0 +1,7 @@
+export type ScheduleData = {
+  id: number
+  from: string
+  to: string
+  depTime: string
+  arrTime: string
+}
