+++ src/types/ship.ts
@@ -1,0 +1,10 @@
+export type ShipData = {
+  id: number
+  regiNumber: string
+};
+
+export type SelectShip = {
+  id?: string
+  selectShip?: string
+
+}
